- Moved old photo tests from UserControllerTest to UserPhotoControllerTest - Moved old photo tests from EventControllerTest to EventPhotoControllerTest
