+++ EventsExpress.Test/ControllerTests/EventControllerTests.cs
@@ -35,15 +35,12 @@ namespace EventsExpress.Test.ControllerTests
         private Mock<IEventService> service;
         private EventController eventController;
         private Mock<ISecurityContext> mockSecurityContextService;
-        private Mock<IPhotoService> mockPhotoservice;
         private Mock<IValidator<IFormFile>> mockValidator;
         private Mock<IValidator<EventViewModel>> mockEventViewModelValidator;
         private UserDto _userDto;
         private Guid _idUser = Guid.NewGuid();
         private Guid _eventId = Guid.NewGuid();
         private string _userEmal = "user@gmail.com";
-
-        public PhotoService PhotoService { get; set; }
 
         private Mock<IMapper> MockMapper { get; set; }
 
@@ -52,56 +49,18 @@ namespace EventsExpress.Test.ControllerTests
         {
             MockMapper = new Mock<IMapper>();
             mockSecurityContextService = new Mock<ISecurityContext>();
-            mockPhotoservice = new Mock<IPhotoService>();
             service = new Mock<IEventService>();
             mockValidator = new Mock<IValidator<IFormFile>>();
             mockEventViewModelValidator = new Mock<IValidator<EventViewModel>>();
-            eventController = new EventController(service.Object, MockMapper.Object, mockSecurityContextService.Object, mockPhotoservice.Object);
+            eventController = new EventController(service.Object, MockMapper.Object, mockSecurityContextService.Object);
             eventController.ControllerContext = new ControllerContext();
             eventController.ControllerContext.HttpContext = new DefaultHttpContext();
-
-            var mockBlobServiceClient = new Mock<BlobServiceClient>();
-            var mockBlobContainer = new Mock<BlobContainerClient>();
-            Mock<BlobClient> blobClientMock = new Mock<BlobClient>();
-            blobClientMock = new Mock<BlobClient>();
-
-            var mockOpt = new Mock<IOptions<ImageOptionsModel>>();
-            mockOpt.Setup(opt => opt.Value.Thumbnail).Returns(400);
-            mockOpt.Setup(opt => opt.Value.Image).Returns(1200);
-
-            mockBlobServiceClient
-                .Setup(s => s.GetBlobContainerClient(It.IsAny<string>()))
-                .Returns(mockBlobContainer.Object);
-
-            mockBlobContainer
-                .Setup(c => c.GetBlobClient(It.IsAny<string>()))
-                .Returns(blobClientMock.Object);
-
-            PhotoService = new PhotoService(mockOpt.Object, mockBlobServiceClient.Object);
 
             _userDto = new UserDto
             {
                 Id = _idUser,
                 Email = _userEmal,
             };
-        }
-
-        [Test]
-        [TestCase(@"./Images/valid-image.jpg")]
-        public void SetEventTempPhoto_ValidPhoto(string testFilePath)
-        {
-            byte[] bytes = File.ReadAllBytes(testFilePath);
-            string base64 = Convert.ToBase64String(bytes);
-            string fileName = Path.GetFileName(testFilePath);
-            using var stream = new MemoryStream(Encoding.UTF8.GetBytes(base64));
-            var file = new FormFile(stream, 0, stream.Length, null, fileName)
-            {
-                Headers = new HeaderDictionary(),
-                ContentType = "application/octet-stream",
-            };
-            Guid id = Guid.NewGuid();
-
-            Assert.DoesNotThrowAsync(async () => await PhotoService.AddEventTempPhoto(file, id));
         }
 
         [Test]
@@ -168,13 +127,6 @@ namespace EventsExpress.Test.ControllerTests
             service.Setup(e => e.Exists(_eventId)).Returns(true);
             var expected = eventController.GetCurrentRate(_eventId);
             Assert.IsInstanceOf<OkObjectResult>(expected);
-        }
-
-        [Test]
-        public void SetEventTempPhoto_OkResult()
-        {
-            var expected = eventController.SetEventTempPhoto(_eventId, new PhotoViewModel());
-            Assert.IsInstanceOf<OkResult>(expected.Result);
         }
 
         [Test]
